establish high-level chatbot server
