+++ app.js
@@ -1,0 +1,26 @@
+var restify = require('restify');
+var builder = require('botbuilder');
+var assessBot = require('./assess-bot');
+
+// var config = require('./config');
+
+//=========================================================
+// Bot Setup
+//=========================================================
+
+// Setup Restify Server
+var server = restify.createServer();
+
+server.listen(process.env.port || process.env.PORT || 3978, function () {
+    console.log('%s listening to %s', server.name, server.url);
+});
+
+// Create chat bot
+var connector = new builder.ChatConnector({
+    appId: process.env.MICROSOFT_APP_ID,
+    appPassword: process.env.MICROSOFT_APP_PASSWORD
+});
+
+assessBot.create(connector);
+
+server.post('/api/messages', connector.listen());
